% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_build.R
\name{corpus_build_author}
\alias{corpus_build_author}
\title{Build a corpus}
\usage{
corpus_build_author(guten_author)
}
\arguments{
\item{guten_author}{A character vector, the author of interest. See
\code{gutenbergr::gutenberg_metadata} for more information.}
}
\value{
A \code{cps_corpus} object.
}
\description{
Construct a corpus from either a list of works or an author.
}
\author{
Charlie Gallagher
}
