% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_build.R
\name{corpus_build_works}
\alias{corpus_build_works}
\title{Build a corpus}
\usage{
corpus_build_works(works)
}
\arguments{
\item{works}{A numeric vector. See \code{gutenbergr::gutenberg_metadata}
for more information.}
}
\value{
A \code{cps_corpus} object.
}
\description{
Construct a corpus from either a list of works or an author.
}
\author{
Charlie Gallagher
}
