% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps.R
\name{cps_bk}
\alias{cps_bk}
\title{Build a \code{cps_bk} object}
\usage{
cps_bk(x, text = "text", author = "author", title = "title")
}
\arguments{
\item{x}{Data frame to be converted to \code{cps_bk}.}

\item{text}{Name of the column which contains the text.}

\item{author}{Name of the column with contains the author's name.}

\item{title}{Name of the column which contains the title.}
}
\value{
A \code{cps_bk}.
}
\description{
Constructor function.
}
\author{
Charlie Gallagher
}
