% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_search.R
\name{cps_bk_search}
\alias{cps_bk_search}
\title{Search a \code{cps_bk}}
\usage{
cps_bk_search(cps_bk, word)
}
\arguments{
\item{cps_bk}{A \code{cps_bk} object.}

\item{word}{A character vector of length 1.}
}
\description{
Search a \code{cps_bk} for instances of a word.
}
\author{
Charlie Gallagher
}
