% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps.R
\name{cps_corpus}
\alias{cps_corpus}
\title{Build a \code{cps_corpus} object}
\usage{
cps_corpus(cps_bk_list)
}
\arguments{
\item{cps_bk_list}{A list of \code{cps_bk} objects.}
}
\value{
A \code{cps_corpus}.
}
\description{
Constructor function for building a \code{cps_corpus} object from a list
of \code{cps_bk} objects. This is the main unit of analysis in the rest
of the package, and even single books are better kept in a \code{cps_corpus}
rather than a \code{cps_bk} for analytical purposes.
}
\author{
Charlie Gallagher
}
