% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guten_grab.R
\name{guten_grab}
\alias{guten_grab}
\title{Gutenberg texts properly formatted}
\usage{
guten_grab(code)
}
\arguments{
\item{code}{Book code from \code{gutenbergr::gutenberg_metadata}. Numeric.}
}
\value{
A \code{cps_bk}.
}
\description{
Get a book's text from Project Gutenberg via \code{gutenbergr}. The formatting
is corrected, so that it returns sentences rather than lines.
}
\details{
This is the details section, supposedly. It may have its own sections.
}
\section{Warning}{

Do not operate heavy machinery within 8 hours of using this function.
}

\seealso{
\code{\link{guten_search}}, \code{\link{corpus_search_author}},
\code{\link{corpus_search_works}}.
}
\author{
Charlie Gallagher
}
