% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guten_search_author.R
\name{guten_search_author}
\alias{guten_search_author}
\title{Search a gutenberg author's complete works for a word.}
\usage{
guten_search_author(guten_author, word)
}
\arguments{
\item{guten_author}{Author's name in format given by
\code{gutenbergr::gutenberg_metadata}. Character.}

\item{word}{Word of interest. Character.}
}
\value{
A \code{cps_corpus} object.
}
\description{
Calls \code{guten_grab} iteratively, then returns sentences containing the
specified word.
}
\author{
Charlie Gallagher
}
