% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guten_search_works.R
\name{guten_search_works}
\alias{guten_search_works}
\title{Search a set of gutenberg works for a word.}
\usage{
guten_search_works(work_ids, word)
}
\arguments{
\item{word}{Word of interest. Character.}

\item{word_ids}{Numeric vector of book IDs
from \code{gutenbergr::gutenberg_metadata}.}
}
\value{
A \code{cps_corpus} object.
}
\description{
Calls \code{guten_grab} iteratively, then returns sentences containing the
specified word.
}
\author{
Charlie Gallagher
}
