% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps.R
\name{print.cps_bk}
\alias{print.cps_bk}
\title{Print a \code{cps_bk}}
\usage{
\method{print}{cps_bk}(x, ..., n_sentences = 10)
}
\arguments{
\item{x}{a \code{cps_bk} object}

\item{...}{Further arguments. In this case, not used.}

\item{n_sentences}{Number of lines to print.}
}
\description{
A print method for \code{cps_bk} objects.
}
\author{
Charlie Gallagher
}
