% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps.R
\name{print.cps_corpus}
\alias{print.cps_corpus}
\title{Print a \code{cps_corpus}}
\usage{
\method{print}{cps_corpus}(x, ...)
}
\arguments{
\item{x}{a \code{cps_corpus} object}

\item{...}{Further arguments. In this case, not used.}
}
\description{
A print method for \code{cps_corpus} objects.
}
\author{
Charlie Gallagher
}
