% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guten_search.R
\name{guten_search}
\alias{guten_search}
\title{Search a gutenbergr text for a word}
\usage{
guten_search(code, word)
}
\arguments{
\item{code}{Book code from \code{gutenbergr::gutenberg_metadata}. Numeric.}

\item{word}{Word of interest. Character.}
}
\value{
A \code{cps_bk}, or NULL if no book is returned from gutenbergr.
}
\description{
Calls \code{guten_grab} and returns sentences containing the specified word.
}
\author{
Charlie Gallagher
}
